+++ CHANGELOG.md
@@ -18,7 +18,9 @@
  - #249 Redfish Events notification
  - #233 Implement ResourceBlocks
  - #234 Implement ResourceZones
+ - #258 Check OneView availability
  - #259 Add ResourceBlocks for each Server Profile Template in the ResourceBlocks collection
+
 
 # New Redfish resources
  - EventService